requirements.txt
# Medical Image Analysis Platform
# Core dependencies for AI-Assisted Medical Image Analysis

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Segment Anything Model (SAM)
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Vision-Language Models
transformers>=4.35.0
accelerate>=0.24.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Medical Imaging (optional, for DICOM support)
pydicom>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web Interface
gradio>=4.0.0

# Evaluation Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0

# Jupyter Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0

# HuggingFace Hub
huggingface-hub>=0.17.0

